change payment amoutn

+++ app/src/main/java/com/example/spacebunsmobileapp/data/Data.kt
@@ -107,7 +107,7 @@ data class DonationEventDetail(
     @DocumentId
     var donationId: String = "",
     var donorName: String = "",
-    var donationAmount: Double = 0.00,
+    var donationAmount: Int = 0,
     var donationEventId: String = "",
     var donationDate: Date = Date() // current Date
 ) {
